# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tasty Crousty - Smart food ordering and delivery platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      summary: List restaurants
      parameters:
        - name: cityId
          in: query
          schema:
            type: ["integer", "null"]
        - name: zoneId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Restaurants list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"
    post:
      operationId: createRestaurant
      tags: [restaurants]
      summary: Create / register a restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRestaurantBody"
      responses:
        "201":
          description: Restaurant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /restaurants/{restaurantId}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      summary: Get restaurant by ID
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
    patch:
      operationId: updateRestaurant
      tags: [restaurants]
      summary: Update restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRestaurantBody"
      responses:
        "200":
          description: Updated restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /restaurants/{restaurantId}/approve:
    post:
      operationId: approveRestaurant
      tags: [restaurants]
      summary: Approve a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /restaurants/{restaurantId}/reject:
    post:
      operationId: rejectRestaurant
      tags: [restaurants]
      summary: Reject a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /restaurants/{restaurantId}/toggle-open:
    post:
      operationId: toggleRestaurantOpen
      tags: [restaurants]
      summary: Toggle restaurant open/closed
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /restaurants/{restaurantId}/hours:
    get:
      operationId: getRestaurantHours
      tags: [restaurants]
      summary: Get restaurant hours
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hours list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestaurantHours"
    post:
      operationId: setRestaurantHours
      tags: [restaurants]
      summary: Set restaurant hours
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetRestaurantHoursBody"
      responses:
        "200":
          description: Hours set
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RestaurantHours"
  /restaurants/{restaurantId}/stats:
    get:
      operationId: getRestaurantStats
      tags: [restaurants]
      summary: Get restaurant performance stats
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestaurantStats"

  # MENUS
...
components:
  schemas:
...
    Restaurant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        cityId:
          type: ["integer", "null"]
        zoneId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, approved, rejected, suspended]
        isOpen:
          type: boolean
        category:
          type: ["string", "null"]
        avgRating:
          type: ["number", "null"]
        totalOrders:
          type: ["integer", "null"]
        estimatedPrepTime:
          type: ["integer", "null"]
        commissionRate:
          type: ["number", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - status
        - isOpen
        - createdAt
    CreateRestaurantBody:
      type: object
      properties:
        name:
          type: string
        nameAr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        cityId:
          type: ["integer", "null"]
        zoneId:
          type: ["integer", "null"]
        category:
          type: ["string", "null"]
        estimatedPrepTime:
          type: ["integer", "null"]
      required:
        - name
    UpdateRestaurantBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        nameAr:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        cityId:
          type: ["integer", "null"]
        zoneId:
          type: ["integer", "null"]
        isOpen:
          type: ["boolean", "null"]
        category:
          type: ["string", "null"]
        estimatedPrepTime:
          type: ["integer", "null"]
        commissionRate:
          type: ["number", "null"]
        status:
          type: ["string", "null"]
    RestaurantHours:
      type: object
      properties:
        id:
          type: integer
        dayOfWeek:
          type: integer
        openTime:
          type: string
        closeTime:
          type: string
        isOpen:
          type: boolean
      required:
        - id
        - dayOfWeek
        - openTime
        - closeTime
        - isOpen
    SetRestaurantHoursBody:
      type: object
      properties:
        hours:
          type: array
          items:
            type: object
            properties:
              dayOfWeek:
                type: integer
              openTime:
                type: string
              closeTime:
                type: string
              isOpen:
                type: boolean
            required:
              - dayOfWeek
              - openTime
              - closeTime
              - isOpen
      required:
        - hours
    RestaurantStats:
      type: object
      properties:
        totalOrders:
          type: integer
        completedOrders:
          type: integer
        cancelledOrders:
          type: integer
        avgRating:
          type: number
        avgPrepTime:
          type: number
        revenue:
          type: number
        cancellationRate:
          type: number
      required:
        - totalOrders
        - completedOrders
        - cancelledOrders
        - avgRating
        - avgPrepTime
        - revenue
        - cancellationRate

    # MENUS
...